Spring REST - Web Client

+++ Spring-11-REST-WebClient/src/main/java/com/restwebflux/controller/WebFluxController.java
@@ -104,7 +104,13 @@ public class WebFluxController {
     }
 
     @DeleteMapping("/delete/{id}")
-    public Mono<Void> deleteWebClient(@PathVariable("id") Long id){
+    public Mono<Void> deleteWebClient(@PathVariable("id") Long id) throws Exception {
+
+        Integer countGenres = genreRepository.countGenresNativeQuery(id);
+        if (countGenres > 0) {
+            throw new Exception("Genre can't be deleted, is linked by a movie");
+        }
+
         return webClient
                 .delete()
                 .uri("/delete-genre/{id}", id)
